requirements.txt
aiocache==0.11.1
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
cycler==0.10.0
discord==1.0.1
discord.py==1.6.0
idna==3.1
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
numpy==1.20.1
Pillow==8.1.0